ValueImpl
